+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Symfony Mailer: Love Sending Emails Again
+## [17.02.2022]
+- Router Request Context: Fix Paths in the CLI
+- Using a Base Email Template
+## [15.02.2022]
+- Let's Make a Console Command!
+- Ink: Automatic CSS Email Framework
 ## [12.02.2022]
 - Inlining CSS Files
 - Automatic CSS Inlining